sidebar and page defaults refactor
