+++ packages/ui/components/DashboardSidebar.tsx
@@ -1,18 +1,6 @@
 import React from "react";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import {
-  faCog,
-  faEnvelope,
-  faUserFriends,
-  faSearch,
-  faCommentAlt,
-} from "@fortawesome/free-solid-svg-icons";
 import Sidebar from "./Sidebar2";
-
-const iconStyle: React.CSSProperties = {
-  width: 15,
-};
-
+import { SIDEBAR_LIST } from "constants/page";
 export interface SidebarItemProp {
   section: string;
   icon: JSX.Element | null;
@@ -21,29 +9,6 @@ export interface SidebarItemProp {
   renderIcon?: (active: boolean) => JSX.Element;
 }
 
-const ITEMS = [
-  {
-    section: "Blasts",
-    icon: <FontAwesomeIcon icon={faCommentAlt} style={iconStyle} />,
-    route: "/blasts",
-  },
-  {
-    section: "Audiences",
-    icon: <FontAwesomeIcon icon={faUserFriends} style={iconStyle} />,
-    route: "/audiences",
-  },
-  {
-    section: "Emails",
-    icon: <FontAwesomeIcon icon={faEnvelope} style={iconStyle} />,
-    route: "/emails",
-  },
-  {
-    section: "Person Explorer",
-    icon: <FontAwesomeIcon icon={faSearch} style={iconStyle} />,
-    route: "/explorer",
-  },
-];
-
 interface Props {
   index: number;
   onClick: ({ index: number, route: string }) => void;
@@ -51,10 +16,10 @@ interface Props {
 
 const DashboardSidebar = (props: Props) => (
   <Sidebar
-    items={ITEMS}
+    items={SIDEBAR_LIST}
     index={props.index}
     onClick={(index: number) => {
-      const route = ITEMS[index].route;
+      const route = SIDEBAR_LIST[index].route;
       props.onClick({ index, route });
     }}
   />
