--- packages/ui/constants/page.ts
@@ -1,33 +1,0 @@
-export const PAGE_DEFAULTS = {
-  explorer: {
-    index: {
-      title: "Person Explorer",
-      subtitle: "Filter and search through your People data.",
-    },
-  },
-  events: {
-    index: {
-      title: "Events Explorer",
-      subtitle: "Filter and search through your Events data.",
-    },
-    id: {
-      title: "",
-      subtitle: "View the data associated with this Event instance.",
-    },
-  },
-  emails: {
-    id: {
-      title: "",
-      placeholder: "Untitled Email",
-      subtitle: "Create an email to send to a list of customers.",
-    },
-  },
-  blasts: {
-    index: {
-      title: "Blasts",
-      subtitle: "View and manage your message blasts.",
-      actionText: "+ Send Blast",
-      actionUrl: "/blasts/create",
-    },
-  },
-};
